apps/ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
kafka-python==2.0.2
grpcio==1.59.0
grpcio-tools==1.59.0
redis==5.0.1
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
tensorflow==2.15.0
tensorflow-recommenders==0.7.3
tensorflow-datasets==4.9.3